bill done

+++ routes/stripe.js
@@ -98,7 +98,7 @@ router.post("/subscribe_plan", async function (req, res, next) {
                         UserMeta.bulkCreate(userMetaList)
                             .then((result) => {
                                 if (result) {
-                                    deleteLawyerFacilityAddons(req.body.userId, function (deleteLawyerFacility) {
+                                    deleteLawyerFacilityAddons(req.body.userId, req.body.type, function (deleteLawyerFacility) {
                                         if (deleteLawyerFacility) {
                                             setLawyerFacilityAddons(
                                                 req.body.facilityList,
@@ -279,7 +279,7 @@ router.post("/update_plan", passport.authenticate("jwt", { session: false }),
                                             })
                                             .then((subscribePlan) => {
                                                 deleteLawyerFacilityAddons(
-                                                    req.body.userId,
+                                                    req.body.userId, req.body.type,
                                                     function (deleteFacilityLawyer) {
                                                         if (deleteFacilityLawyer) {
                                                             setLawyerFacilityAddons(
@@ -314,49 +314,59 @@ router.post("/update_plan", passport.authenticate("jwt", { session: false }),
                             })
                             .catch(next);
                     })
-                    .catch(next => console.log(next));
+                    .catch(next);
             })
         });
     }
 );
 
-function deleteLawyerFacilityAddons(userId, callback) {
-    Lawyer_Facility.findAll({ where: { lawyerId: userId } }).then(
-        (lawyerFacility) => {
-            if (lawyerFacility && lawyerFacility.length > 0) {
-                let count = 0;
-                lawyerFacility.forEach((element, index, Array) => {
-                    Lawyer_Facility.destroy({
-                        where: { lawyer_facilityId: element.lawyer_facilityId },
-                    }).then((data) => {
-                        if (count === Array.length - 1) {
-                            callback(data);
-                        }
-                        count++;
+function deleteLawyerFacilityAddons(userId, type, callback) {
+    console.log(userId)
+    if (type == 'lawyer') {
+        Lawyer_Facility.findAll({ where: { lawyerId: userId } }).then(
+            (lawyerFacility) => {
+
+                if (lawyerFacility && lawyerFacility.length > 0) {
+                    let count = 0;
+                    lawyerFacility.forEach((element, index, Array) => {
+                        Lawyer_Facility.destroy({
+                            where: { lawyer_facilityId: element.lawyer_facilityId },
+                        }).then((data) => {
+                            if (count === Array.length - 1) {
+                                callback(data);
+                            }
+                            count++;
+                        });
                     });
-                });
-            } else {
-                Defender_Facility.findAll({ where: { defenderId: userId } }).then(
-                    (defenderFacility) => {
-                        if (defenderFacility && defenderFacility.length > 0) {
-                            let count = 0;
-                            defenderFacility.forEach((element, index, Array) => {
-                                Defender_Facility.destroy({
-                                    where: { defender_facilityId: element.defender_facilityId },
-                                }).then((data) => {
-                                    if (count === Array.length - 1) {
-                                        callback(data);
-                                    }
-                                    count++;
-                                });
-                            });
-                        }
-                        else {
-                            callback(true);
-                        }
+                }
+                else {
+                    callback(true);
+                }
+
+            }).catch(next)
+    }
+    if (type == 'defender') {
+        Defender_Facility.findAll({ where: { defenderId: userId } }).then(
+            (defenderFacility) => {
+                if (defenderFacility && defenderFacility.length > 0) {
+                    let count = 0;
+                    defenderFacility.forEach((element, index, Array) => {
+                        Defender_Facility.destroy({
+                            where: { defender_facilityId: element.defender_facilityId },
+                        }).then((data) => {
+                            if (count === Array.length - 1) {
+                                callback(data);
+                            }
+                            count++;
+                        });
                     });
-            }
-        })
+                }
+                else {
+                    callback(true);
+                }
+            });
+    }
+
 }
 
 function setLawyerFacilityAddons(facilityList, callback) {
@@ -378,7 +388,7 @@ function setLawyerFacilityAddons(facilityList, callback) {
                         callback(result);
                     }
                     count++;
-                })
+                }).catch(next)
         }
     });
 }
